# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universe Moon [Um] API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  token: string;
  username: string;
  password: string;
  avatarUrl?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
  avatarUrl?: string;
  xp?: number;
  level?: number;
  isBanned?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  success: boolean;
  user?: User;
  message?: string;
}

export interface TokenValidationResponse {
  valid: boolean;
  role?: string;
}

export interface Member {
  id: number;
  name: string;
  nickname: string;
  role?: string;
  bio?: string;
  joinDate?: string;
  favoriteSong?: string;
  socialLinks?: string;
  avatarUrl?: string;
  isActive: boolean;
  kickReason?: string;
  kickDate?: string;
}

export interface CreateMemberRequest {
  name: string;
  nickname: string;
  role?: string;
  bio?: string;
  joinDate?: string;
  favoriteSong?: string;
  socialLinks?: string;
  avatarUrl?: string;
}

export interface UpdateMemberRequest {
  name?: string;
  nickname?: string;
  role?: string;
  bio?: string;
  favoriteSong?: string;
  socialLinks?: string;
  avatarUrl?: string;
}

export type ChatMessageReactions = { [key: string]: number };

export interface ChatMessage {
  id: number;
  sender: string;
  content: string;
  replyToId?: number;
  replyToContent?: string;
  isSticker?: boolean;
  stickerCode?: string;
  createdAt: string;
  reactions?: ChatMessageReactions;
}

export interface SendChatMessageRequest {
  sender: string;
  content: string;
  replyToId?: number;
  isSticker?: boolean;
  stickerCode?: string;
}

export type NglMessageReactions = { [key: string]: number };

export interface NglMessage {
  id: number;
  content: string;
  recipient?: string;
  createdAt: string;
  reactions?: NglMessageReactions;
}

export interface SendNglRequest {
  content: string;
  recipient?: string;
}

export interface Photo {
  id: number;
  url: string;
  caption?: string;
  uploadedBy?: string;
  category?: string;
  createdAt: string;
}

export interface AddPhotoRequest {
  url: string;
  caption?: string;
  uploadedBy?: string;
  category?: string;
}

export interface Memory {
  id: number;
  title: string;
  content: string;
  date?: string;
  addedBy?: string;
  photo?: string;
  createdAt: string;
}

export interface AddMemoryRequest {
  title: string;
  content: string;
  date?: string;
  addedBy?: string;
  photo?: string;
}

export interface Link {
  id: number;
  title: string;
  url: string;
  desc?: string;
  category?: string;
  addedBy?: string;
  createdAt: string;
}

export interface AddLinkRequest {
  title: string;
  url: string;
  desc?: string;
  category?: string;
  addedBy?: string;
}

export interface Track {
  id: number;
  title: string;
  artist: string;
  url?: string;
  addedBy?: string;
  createdAt: string;
}

export interface AddTrackRequest {
  title: string;
  artist: string;
  url?: string;
  addedBy?: string;
}

export interface Secret {
  id: number;
  content: string;
  revealDate?: string;
  locked: boolean;
  createdAt: string;
}

export interface AddSecretRequest {
  content: string;
  revealDate?: string;
}

export interface OpMem {
  id: number;
  title: string;
  description?: string;
  openDate?: string;
  closeDate?: string;
  isOpen: boolean;
  acceptedMembers?: string[];
  createdAt: string;
}

export interface CreateOpMemRequest {
  title: string;
  description?: string;
  openDate?: string;
  closeDate?: string;
  isOpen?: boolean;
}

export interface Mep {
  id: number;
  title: string;
  description?: string;
  videoUrl: string;
  participants?: string[];
  createdBy?: string;
  createdAt: string;
}

export interface CreateMepRequest {
  title: string;
  description?: string;
  videoUrl: string;
  participants?: string[];
  createdBy?: string;
}

export interface GamePlayer {
  username: string;
  role?: string;
  isAlive: boolean;
  isRevealed?: boolean;
}

export interface GameState {
  id: number;
  gameType: string;
  status: string;
  players: GamePlayer[];
  phase: string;
  round?: number;
  winner?: string;
  createdAt?: string;
}

export interface CreateGameRequest {
  gameType: string;
  players: string[];
}

export interface GameActionRequest {
  action: string;
  targetUsername?: string;
  actorUsername: string;
}

export interface LeaderboardEntry {
  id: number;
  username: string;
  avatarUrl?: string;
  xp: number;
  level: number;
  rank: number;
}

export interface AddReactionRequest {
  contentType: string;
  contentId: number;
  emoji: string;
  username: string;
}

export type VoteResultVotes = { [key: string]: number };

export interface VoteResult {
  topic: string;
  votes?: VoteResultVotes;
  totalVotes: number;
  userVote?: string;
}

export interface CastVoteRequest {
  topic: string;
  option: string;
  username: string;
}

export type ValidateTokenBody = {
  token: string;
};

export type KickMemberBody = {
  reason: string;
};

export type GetChatMessagesParams = {
  limit?: number;
};

export type GetPhotosParams = {
  category?: string;
};

export type AddAcceptedMemberBody = {
  name: string;
  opMemId: number;
};

export type GetVotesParams = {
  topic?: string;
};

export type ResetVotesBody = {
  topic: string;
};