# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Universe Moon [Um] API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/validate-token:
    post:
      operationId: validateToken
      tags: [auth]
      summary: Validate registration token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenValidationResponse"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/{id}:
    delete:
      operationId: deleteChatMessage
      tags: [chat]
      summary: Delete chat message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        password:
          type: string
        avatarUrl:
          type: string
      required: [token, username, password]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [success]
    TokenValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        role:
          type: string
      required: [valid]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
        xp:
          type: integer
        level:
          type: integer
        isBanned:
          type: boolean
        createdAt:
          type: string
      required: [id, username, role]
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        content:
          type: string
        replyToId:
          type: integer
        replyToContent:
          type: string
        isSticker:
          type: boolean
        stickerCode:
          type: string
        createdAt:
          type: string
        reactions:
          type: object
          additionalProperties:
            type: integer
      required: [id, sender, content, createdAt]
    SendChatMessageRequest:
      type: object
      properties:
        sender:
          type: string
        content:
          type: string
        replyToId:
          type: integer
        isSticker:
          type: boolean
        stickerCode:
          type: string
      required: [sender, content]
...