# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universe Moon [Um] API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  role: string;
  avatarUrl?: string;
  xp?: number;
  level?: number;
  isBanned?: boolean;
  createdAt?: string;
}